+++ shared/union.go
@@ -1,0 +1,33 @@
+// File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
+
+package shared
+
+type UnionString string
+
+func (UnionString) ImplementsAggregationCustomFieldsUnion()                     {}
+func (UnionString) ImplementsAggregationNewParamsCustomFieldsUnion()            {}
+func (UnionString) ImplementsAggregationUpdateParamsCustomFieldsUnion()         {}
+func (UnionString) ImplementsCompoundAggregationCustomFieldsUnion()             {}
+func (UnionString) ImplementsCompoundAggregationNewParamsCustomFieldsUnion()    {}
+func (UnionString) ImplementsCompoundAggregationUpdateParamsCustomFieldsUnion() {}
+func (UnionString) ImplementsMeterCustomFieldsUnion()                           {}
+func (UnionString) ImplementsMeterNewParamsCustomFieldsUnion()                  {}
+func (UnionString) ImplementsMeterUpdateParamsCustomFieldsUnion()               {}
+func (UnionString) ImplementsProductCustomFieldsUnion()                         {}
+func (UnionString) ImplementsProductNewParamsCustomFieldsUnion()                {}
+func (UnionString) ImplementsProductUpdateParamsCustomFieldsUnion()             {}
+
+type UnionFloat float64
+
+func (UnionFloat) ImplementsAggregationCustomFieldsUnion()                     {}
+func (UnionFloat) ImplementsAggregationNewParamsCustomFieldsUnion()            {}
+func (UnionFloat) ImplementsAggregationUpdateParamsCustomFieldsUnion()         {}
+func (UnionFloat) ImplementsCompoundAggregationCustomFieldsUnion()             {}
+func (UnionFloat) ImplementsCompoundAggregationNewParamsCustomFieldsUnion()    {}
+func (UnionFloat) ImplementsCompoundAggregationUpdateParamsCustomFieldsUnion() {}
+func (UnionFloat) ImplementsMeterCustomFieldsUnion()                           {}
+func (UnionFloat) ImplementsMeterNewParamsCustomFieldsUnion()                  {}
+func (UnionFloat) ImplementsMeterUpdateParamsCustomFieldsUnion()               {}
+func (UnionFloat) ImplementsProductCustomFieldsUnion()                         {}
+func (UnionFloat) ImplementsProductNewParamsCustomFieldsUnion()                {}
+func (UnionFloat) ImplementsProductUpdateParamsCustomFieldsUnion()             {}